fix bug; add license header

+++ pytidbrep/binlogstream.py
@@ -1,4 +1,21 @@
 # -*- coding: utf-8 -*-
+
+# Copyright (c) 2016 HashData Inc.
+# All Rights Reserved.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+
+
 from __future__ import unicode_literals
 
 import logging
@@ -46,7 +63,7 @@ class BinLogStreamReader(object):
         self._current_binlog_file = None
         self._current_binlog_file_name = None
         self._current_binlog_file_position = 0
-        self._skip_to_timestamp = skip_to_timestamp
+        self._skip_to_timestamp = long(skip_to_timestamp)
         self._blocking = blocking
         self._ignore_error = ignore_error
         self._current_events = []
